Now make check passes also when some of the back-ends is disabled.
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@3468 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/conjunct-functionality.sh
@@ -1,13 +1,15 @@
 #!/bin/sh
 TOOLDIR=../../tools/src
 for i in "" .sfst .ofst .foma; do
-    if test -f cat_or_dog$i -a -f cat$i ; then
-        if ! $TOOLDIR/hfst-conjunct cat_or_dog$i cat$i > test ; then
-            exit 1
+    if ((test -z "$i") || $TOOLDIR/hfst-format --list-formats | grep $i > /dev/null); then
+        if test -f cat_or_dog$i -a -f cat$i ; then
+            if ! $TOOLDIR/hfst-conjunct cat_or_dog$i cat$i > test ; then
+                exit 1
+            fi
+            if ! $TOOLDIR/hfst-compare -s test cat$i  ; then
+                exit 1
+            fi
+            rm test;
         fi
-        if ! $TOOLDIR/hfst-compare -s test cat$i  ; then
-            exit 1
-        fi
-        rm test;
     fi
 done
